introduce standard-config..go
